Extract agent chat action gate
